bots/GroupBot/requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2023.11.17
python-telegram-bot==13.7
pytz==2023.3.post1
six==1.16.0
tornado==6.3.3
tzlocal==5.2
